refactor: 함수 분리 및 형 변환 방법 변경
